+++ Move.h
@@ -8,7 +8,7 @@ typedef uint16_t move;
 
 namespace Move
 {
-    move build_move(const uint8_t s, const uint8_t e, const uint8_t p = 0);
+    move build_move(const uint16_t s, const uint16_t e, const uint16_t p = 0);
     uint8_t start(const move& m);
     uint8_t end(const move& m);
     uint8_t promote(const move& m);
